[BugFix] Fix on forward new message
Fixed a bug where on new forward message, didn't append to the list of messaegs if thread was opnned by forward list and forward a message to the thread. It is due to ThreadsViewModel.threads is not available but in Forward picker it is available so it ignored message insertion by default.