[BugFix] Fix selected background color after reconnect
If the user stays at a thread and there is a reconnect there without leaving the thread, after new conversations arrive it will lead to create new object for new conversations. So, SwiftUI will deallocate it's publisher/subscriber and the ThreadListBackground won't be triggered by the SwiftUI. The best possible solution for this matter is that deselct the row before releaseing the original object.