[BugFix] Fix last message scroll position
Now only the deceleration methods for scrollView can detect if we are about to show the last message or not.
By utilizing this methods not only we will reduce the number of comparisions in willDidplayCell but also we will get more accurate result.
Also once the user hit move to buttom button, we are sure that isAtBottomOfTheList should be set to true. And we have to do this
because once we move programatically table view delegates for scrollview won't be called.
We also fixed contentInset to be set at didLayoutSubviews because it is way more accurte in size of the views and it will rapidly called if there is any change.
We also prevent set contentInset to something when keyboard is showing, because keyboard apperance will take care of this