[BugFix] Fix fixLastMessageIfNeeded
Now we will fix lastMessage of the thread in every situation if the lastMessage itself is not present in the vms we receive in the response.
With this trick we are sure that the message is not there and we have to fix last message.
Also in this commit we have removed isDeletingLastMessage which we weren't use
Also in this version we have fixed the last message delete and edit in the chat SDK that led to incorrect last message in the forward thread list