[BugFix] Fix showing archive threads inside the ThreadsViewModel list
Now in GetThreadRequester we don't do anything with archive and its the caller responsibility to filter the archives thread.
It will lead to fetch archives and fill the cashe however, it won't show archives in the Normal Threads List, and also will show them inside the ThreadOrContact picker where we need to show even the archive threads.