[Refactor] Refactor GalleryView and move image to the view model instead of the SwiftUI View to prevent reloading and blinking image
Now we will only fetch if we are not fething image from the disk, and add a loading progress while we are fetching the image from the disk.
Also in this version we add some loggin information related to fetching image process in the SwiftUI view.
Also in this version we wait to fetch both left and right messages and then combine messages and update the UI once.
Also in this version we refactored the GetHistoryRequester to be able to fetch pure messages beside the MessageRowViewModels