[BugFix] Fix a crash once scrollToBottom method called wrong messageId that was deleted and lead to an indexpath crash
[Refactor] Now most of the lastMessageVO calls are going through a methos in ThreadHistoryViewModel

[BugFix] In this version as a result of a crash we will clear saved scroll position after a disconnect,
and we also prevent saving it if has ever disconnected, unless the user close the thread.

In future commits we will remove more acess to lastMessageVO and make it more centralized.

Also in this version we have removed, an unused method to detect lastDeleted message inside more top,
which was for previous releases.